badge + fixes

+++ lib/generators/lycan_ui/add_generator.rb
@@ -9,15 +9,31 @@ module LycanUi
 
       def copy_views
         copy_file("views/_#{component}.html.erb", "app/views/ui/_#{component}.html.erb")
-        directory("views/#{component}", "app/views/ui/#{component}") if Dir.exist?("app/views/ui/#{component}")
+        directory("views/#{component}", "app/views/ui/#{component}") if dir_exists?("views/#{component}")
       end
 
       def copy_js
-        copy_file("javascript/#{component}_controller.js", "app/javascript/controllers/ui/#{component}_controller.js")
+        if file_exists?("javascript/#{component}_controller.js")
+          copy_file("javascript/#{component}_controller.js", "app/javascript/controllers/ui/#{component}_controller.js")
+        end
 
-        if Dir.exist?("app/javascript/controllers/ui/#{component}")
+        if dir_exists?("javascript/#{component}")
           directory("javascript/#{component}", "app/javascript/controllers/ui/#{component}")
         end
+      end
+
+      private
+
+      def dir_exists?(name)
+        root = source_paths.first
+
+        Dir.exist?("#{root}/#{name}")
+      end
+
+      def file_exists?(name)
+        root = source_paths.first
+
+        File.exist?("#{root}/#{name}")
       end
     end
   end
